+++ js_modules/dagit/packages/app/src/index.tsx
@@ -4,9 +4,11 @@ import './publicPath';
 import {Colors, Icon} from '@blueprintjs/core';
 import {App} from '@dagit/core/app/App';
 import {createAppCache} from '@dagit/core/app/AppCache';
+import {errorLink} from '@dagit/core/app/AppError';
 import {AppProvider} from '@dagit/core/app/AppProvider';
 import {AppTopNav} from '@dagit/core/app/AppTopNav';
 import {ContentRoot} from '@dagit/core/app/ContentRoot';
+import {logLink, timeStartLink} from '@dagit/core/app/apolloLinks';
 import * as React from 'react';
 import ReactDOM from 'react-dom';
 import {Link} from 'react-router-dom';
@@ -16,7 +18,10 @@ import {extractPathPrefix} from './extractPathPrefix';
 
 const pathPrefix = extractPathPrefix();
 
+const apolloLinks = [logLink, errorLink, timeStartLink];
+
 const config = {
+  apolloLinks,
   basePath: pathPrefix,
   origin: process.env.REACT_APP_BACKEND_ORIGIN || document.location.origin,
 };
